feat: 투두리스트 수정 api 구현 #11

+++ src/main/java/itstime/reflog/todolist/repository/TodolistRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import itstime.reflog.member.domain.Member;
 import itstime.reflog.todolist.domain.Todolist;
 
-public interface TodolistRepository extends JpaRepository<Todolist, Integer> {
+public interface TodolistRepository extends JpaRepository<Todolist, Long> {
 
 	List<Todolist> findByMemberAndCreatedDate(Member member, LocalDate date);
 }
